# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: VICIdial Agent UI API
  version: 1.1.0
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/agent/manual2/disposiciones:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: campaignId, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/agent/manual2/subtipificaciones:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: campaignId, in: query, required: false, schema: { type: string } }
        - { name: tipificacion, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/agent/manual2/contacto:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: phoneNumber, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/agent/manual2/formulario:
    post:
      deprecated: true
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/agent/manual2/gestion:
    post:
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/agent/manual2/historial:
    get:
      deprecated: true
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: phoneNumber, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/agent/validacion-claro-peru/disposiciones:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: campaignId, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/agent/validacion-claro-peru/subtipificaciones:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: campaignId, in: query, required: false, schema: { type: string } }
        - { name: tipificacion, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/agent/validacion-claro-peru/formulario:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: documento, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/agent/validacion-claro-peru/gestion:
    post:
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /api/admin/manual2/reporte:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: from, in: query, required: true, schema: { type: string, format: date } }
        - { name: to, in: query, required: true, schema: { type: string, format: date } }
        - { name: campana, in: query, required: false, schema: { type: string } }
        - { name: agente, in: query, required: false, schema: { type: string } }
        - { name: tipificacion, in: query, required: false, schema: { type: string } }
        - { name: disposicion, in: query, required: false, schema: { type: string } }
        - { name: subtipificacion, in: query, required: false, schema: { type: string } }
        - { name: telefono, in: query, required: false, schema: { type: string } }
        - { name: cliente, in: query, required: false, schema: { type: string } }
        - { name: page, in: query, required: false, schema: { type: integer } }
        - { name: size, in: query, required: false, schema: { type: integer } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/admin/manual2/reporte.csv:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: from, in: query, required: true, schema: { type: string, format: date } }
        - { name: to, in: query, required: true, schema: { type: string, format: date } }
        - { name: campana, in: query, required: false, schema: { type: string } }
        - { name: agente, in: query, required: false, schema: { type: string } }
        - { name: tipificacion, in: query, required: false, schema: { type: string } }
        - { name: disposicion, in: query, required: false, schema: { type: string } }
        - { name: subtipificacion, in: query, required: false, schema: { type: string } }
        - { name: telefono, in: query, required: false, schema: { type: string } }
        - { name: cliente, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          description: csv
  /api/admin/manual2/historial:
    get:
      deprecated: true
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: phoneNumber, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/admin/validacion-claro-peru/reporte:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: from, in: query, required: true, schema: { type: string, format: date } }
        - { name: to, in: query, required: true, schema: { type: string, format: date } }
        - { name: campana, in: query, required: false, schema: { type: string } }
        - { name: agente, in: query, required: false, schema: { type: string } }
        - { name: tipificacion, in: query, required: false, schema: { type: string } }
        - { name: disposicion, in: query, required: false, schema: { type: string } }
        - { name: subtipificacion, in: query, required: false, schema: { type: string } }
        - { name: telefono, in: query, required: false, schema: { type: string } }
        - { name: documento, in: query, required: false, schema: { type: string } }
        - { name: encuesta, in: query, required: false, schema: { type: string } }
        - { name: page, in: query, required: false, schema: { type: integer } }
        - { name: size, in: query, required: false, schema: { type: integer } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/admin/validacion-claro-peru/reporte.csv:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: from, in: query, required: true, schema: { type: string, format: date } }
        - { name: to, in: query, required: true, schema: { type: string, format: date } }
        - { name: campana, in: query, required: false, schema: { type: string } }
        - { name: agente, in: query, required: false, schema: { type: string } }
        - { name: tipificacion, in: query, required: false, schema: { type: string } }
        - { name: disposicion, in: query, required: false, schema: { type: string } }
        - { name: subtipificacion, in: query, required: false, schema: { type: string } }
        - { name: telefono, in: query, required: false, schema: { type: string } }
        - { name: documento, in: query, required: false, schema: { type: string } }
        - { name: encuesta, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          description: csv
  /api/admin/manual2/subtipificaciones:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: campana, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    post:
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/admin/manual2/subtipificaciones/{codigo}/activo:
    patch:
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: codigo, in: path, required: true, schema: { type: string } }
        - { name: campana, in: query, required: false, schema: { type: string } }
        - { name: activo, in: query, required: true, schema: { type: boolean } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...